requirements.txt
Click==7.0
Flask==1.0.2
Flask-SocketIO==3.1.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pyserial==3.4
python-engineio==3.7.0
python-socketio==4.0.3
six==1.12.0
Werkzeug==0.15.4
